Released changes for multiple token sign ins and token support in securityConfig
