+++ pnnl.goss.core/goss-client.bnd
@@ -1,3 +1,3 @@
 Private-Package:  \
 	pnnl.goss.core.client
-Bundle-Version: 2.0.172.${tstamp}
+Bundle-Version: 2.0.173.${tstamp}
