+++ pnnl.goss.core/goss-core-commands.bnd
@@ -1,3 +1,3 @@
 Private-Package:  \
 	pnnl.goss.core.commands
-Bundle-Version: 2.0.113.${tstamp}
+Bundle-Version: 2.0.114.${tstamp}
