+++ pnnl.goss.core/goss-core-server-api.bnd
@@ -1,3 +1,3 @@
 Export-Package:  \
 	pnnl.goss.core.server
-Bundle-Version: 3.0.131.${tstamp}
+Bundle-Version: 3.0.132.${tstamp}
