+++ pnnl.goss.core/goss-core-server-registry.bnd
@@ -1,4 +1,4 @@
-Bundle-Version: 1.0.178.${tstamp}
+Bundle-Version: 1.0.179.${tstamp}
 Private-Package:  \
 	pnnl.goss.server.registry
 DynamicImport-Package: *
