+++ pnnl.goss.core/goss-core-server.bnd
@@ -3,4 +3,4 @@ Private-Package:  \
 DynamicImport-Package: *
 #Include-Resource: \
 #	OSGI-INF/blueprint/blueprint.xml=config/blueprint.xml
-Bundle-Version: 2.0.196.${tstamp}
+Bundle-Version: 2.0.197.${tstamp}
