+++ pnnl.goss.core/security-jwt.bnd
@@ -1,2 +1,2 @@
 Private-Package: pnnl.goss.core.security.jwt
-Bundle-Version: 1.0.143.${tstamp}
+Bundle-Version: 1.0.144.${tstamp}
