+++ pnnl.goss.core/security-system.bnd
@@ -1,2 +1,2 @@
 Private-Package: pnnl.goss.core.security.system
-Bundle-Version: 2.0.138.${tstamp}
+Bundle-Version: 2.0.139.${tstamp}
