+++ pnnl.goss.core/src/pnnl/goss/core/packageinfo
@@ -1,1 +1,1 @@
-version 5.1.0
+version 6.0.0
