+++ pnnl.goss.core/src/pnnl/goss/core/security/impl/SecurityConfigImpl.java
@@ -1,16 +1,33 @@
 package pnnl.goss.core.security.impl;
 
 import java.security.SecureRandom;
+import java.text.ParseException;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.Dictionary;
+import java.util.Set;
+import java.util.UUID;
 
 import org.apache.felix.dm.annotation.api.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.nimbusds.jose.JOSEException;
+import com.nimbusds.jose.JWSAlgorithm;
+import com.nimbusds.jose.JWSHeader;
+import com.nimbusds.jose.JWSObject;
+import com.nimbusds.jose.JWSSigner;
+import com.nimbusds.jose.JWSVerifier;
+import com.nimbusds.jose.Payload;
+import com.nimbusds.jose.crypto.MACSigner;
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
 import com.northconcepts.exception.SystemException;
 
 import pnnl.goss.core.security.SecurityConfig;
 import pnnl.goss.core.security.SecurityConstants;
+import pnnl.goss.core.security.jwt.JWTAuthenticationToken;
+import pnnl.goss.core.security.jwt.MACVerifierExtended;
 
 
 @Component
@@ -19,16 +36,25 @@ public class SecurityConfigImpl implements SecurityConfig {
 	private String managerUser;
 	private String managerPassword;
 	private boolean useToken = false;
-	private byte[] tokenSecret = generateSharedKey();
-	
+	private byte[] sharedKey = generateSharedKey();
+
 	private Dictionary<String, Object> properties;
 	private static final Logger log = LoggerFactory.getLogger(SecurityConfigImpl.class);
+	private static final String ISSUED_BY = "GridOPTICS Software System";
 
+	
+	
 	
 	public SecurityConfigImpl(){
 	}
 	
 	
+	protected long getExpirationDate() {
+        return 1000 * 60 * 60 * 24 * 5;
+    }
+
+    protected String getIssuer(){return ISSUED_BY;}
+
 	
 	void updated(Dictionary<String, Object> properties) {
         if (properties != null) {
@@ -42,7 +68,7 @@ public class SecurityConfigImpl implements SecurityConfig {
 			
 			String secret = getProperty(SecurityConstants.PROP_SYSTEM_TOKEN_SECRET, null);
 			if(secret!=null && secret.trim().length()>0){
-				this.tokenSecret = secret.getBytes();
+				this.sharedKey = secret.getBytes();
 			}
 			
 			String useTokenString = getProperty(SecurityConstants.PROP_SYSTEM_USE_TOKEN
@@ -100,12 +126,6 @@ public class SecurityConfigImpl implements SecurityConfig {
 	}
 
 
-
-	@Override
-	public byte[] getTokenSecret() {
-		return tokenSecret;
-	}
-	
 	
     private byte[] generateSharedKey() {
         SecureRandom random = new SecureRandom();
@@ -113,5 +133,76 @@ public class SecurityConfigImpl implements SecurityConfig {
         random.nextBytes(sharedKey);
         return sharedKey;
     }
+    
+    private byte[] getSharedKey(){
+    	if (sharedKey==null )
+    		sharedKey = generateSharedKey();
+		return sharedKey;
+	}
+    
+    public boolean validateToken(String token) {
+    	log.debug("Validate token "+token);
+        try {
+            SignedJWT signed = SignedJWT.parse(token);
+            JWSVerifier verifier = new MACVerifierExtended(getSharedKey(), signed.getJWTClaimsSet());
+            boolean verified = signed.verify(verifier);
+            log.debug("Verified: "+verified);
+            return verified;
+        } catch (ParseException ex) {
+            return false;
+        } catch (JOSEException ex) {
+            return false;
+        }
 
+    }
+    
+    public String createToken(Object userId,  Set<String> roles) {
+    	log.info("Creating token for user "+userId);
+        try {
+        	//TODO, should also include roles(permissions)
+        	
+            JWTClaimsSet.Builder builder = new JWTClaimsSet.Builder();
+            builder.issuer(getIssuer());
+            builder.subject(userId.toString());
+            builder.issueTime(new Date());
+            builder.notBeforeTime(new Date());
+            builder.expirationTime(new Date(new Date().getTime() + getExpirationDate()));
+            builder.jwtID(UUID.randomUUID().toString());
+            
+            JWTAuthenticationToken tokenObj = new JWTAuthenticationToken();
+            tokenObj.setIss(getIssuer());
+            tokenObj.setSub(userId.toString());
+            tokenObj.setIat(new Date().getTime());
+            tokenObj.setNbf(new Date().getTime());
+            tokenObj.setExp(new Date(new Date().getTime() + getExpirationDate()).getTime());
+            tokenObj.setJti(UUID.randomUUID().toString());
+            tokenObj.setRoles(new ArrayList<String>(roles));
+            Payload payload = new Payload(tokenObj.toString());
+
+//            JWTClaimsSet claimsSet = builder.build();
+            JWSHeader header = new JWSHeader(JWSAlgorithm.HS256);
+            JWSObject jwsObject = new JWSObject(header, payload);
+
+            JWSSigner signer = new MACSigner(getSharedKey());
+            jwsObject.sign(signer);
+            return jwsObject.serialize();
+        } catch (JOSEException ex) {
+            return null;
+        }
+    }
+    
+    public JWTAuthenticationToken parseToken(String token){
+    	try{
+	    	SignedJWT signed = SignedJWT.parse(token);
+			Payload payload = signed.getPayload();
+			String jsonToken = payload.toJSONObject().toJSONString();
+			log.info("Json token: "+jsonToken);
+			// look up permissions based on roles and add them
+			JWTAuthenticationToken tokenObj = JWTAuthenticationToken.parse(jsonToken);
+			return tokenObj;
+    	}catch (ParseException e) {
+			// TODO: handle exception
+    		return null;
+		}
+    }
 }
